+++ src/main/java/com/yuuko/core/commands/audio/commands/LoopCommand.java
@@ -18,11 +18,11 @@ public class LoopCommand extends Command {
 
     @Override
     public void onCommand(MessageEvent e) {
-        GuildAudioManager manager = AudioManagerController.getGuildAudioManager(e.getGuild().getId());
+        GuildAudioManager manager = AudioManagerController.getGuildAudioManager(e.getGuild());
 
-        manager.scheduler.setLooping(!manager.scheduler.isLooping());
+        manager.getScheduler().setLooping(!manager.getScheduler().isLooping());
 
-        EmbedBuilder embed = new EmbedBuilder().setTitle("Loop").setDescription("Looping for queue set to **" + manager.scheduler.isLooping() + "**");
+        EmbedBuilder embed = new EmbedBuilder().setTitle("Loop").setDescription("Looping for queue set to **" + manager.getScheduler().isLooping() + "**");
         MessageHandler.sendMessage(e, embed.build());
     }