+++ src/main/java/com/yuuko/core/commands/audio/commands/StopCommand.java
@@ -9,7 +9,7 @@ import com.yuuko.core.events.extensions.MessageEvent;
 import com.yuuko.core.utilities.LavalinkUtilities;
 import lavalink.client.io.Link;
 import net.dv8tion.jda.core.EmbedBuilder;
-import net.dv8tion.jda.core.events.guild.GenericGuildEvent;
+import net.dv8tion.jda.core.entities.Guild;
 
 import java.util.Arrays;
 
@@ -22,9 +22,9 @@ public class StopCommand extends Command {
     @Override
     public void onCommand(MessageEvent e) {
         if(!LavalinkUtilities.isState(e.getGuild(), Link.State.NOT_CONNECTED)) {
-            Configuration.LAVALINK.resetPlayer(e.getGuild().getId());
+            Configuration.LAVALINK.resetPlayer(e.getGuild());
             Configuration.LAVALINK.closeConnection(e.getGuild());
-            AudioManagerController.removeGuildAudioManager(e.getGuild().getId());
+            AudioManagerController.removeGuildAudioManager(e.getGuild());
 
             if(e.getCommand() != null) {
                 EmbedBuilder embed = new EmbedBuilder().setTitle("Stopping").setDescription("Audio connection closed.");
@@ -34,13 +34,13 @@ public class StopCommand extends Command {
     }
 
     /**
-     * Executes command when everyone leaves the channel the bot is in.
+     * Executes command just by feeding the method a guild object.
      *
-     * @param e GenericGuildEvent
+     * @param guild Guild
      */
-    public void onCommand(GenericGuildEvent e) {
-        Configuration.LAVALINK.resetPlayer(e.getGuild().getId());
-        Configuration.LAVALINK.closeConnection(e.getGuild());
-        AudioManagerController.removeGuildAudioManager(e.getGuild().getId());
+    public void onCommand(Guild guild) {
+        Configuration.LAVALINK.resetPlayer(guild);
+        Configuration.LAVALINK.closeConnection(guild);
+        AudioManagerController.removeGuildAudioManager(guild);
     }
 }