+++ src/main/java/com/yuuko/core/scheduler/ScheduleHandler.java
@@ -14,4 +14,8 @@ public class ScheduleHandler {
         tasks.add(schedulerService.scheduleAtFixedRate(job, job.getDelay(), job.getPeriod(), job.getUnit()));
     }
 
+    public static ScheduledFuture registerUniqueJob(Job job) {
+        return schedulerService.schedule(job, job.getDelay(), job.getUnit());
+    }
+
 }