+++ src/main/java/com/yuuko/core/scheduler/Task.java
@@ -2,5 +2,5 @@ package com.yuuko.core.scheduler;
 
 public interface Task {
 
-    void handle();
+    void run();
 }