+++ src/main/java/com/yuuko/core/scheduler/tasks/UpdateMetricsTask.java
@@ -9,7 +9,7 @@ public class UpdateMetricsTask implements Task {
     private static long lastUpdated = 0;
 
     @Override
-    public void handle() {
+    public void run() {
         if(System.currentTimeMillis() - lastUpdated > 29999) {
             MetricsManager.getDiscordMetrics().PING.set(Configuration.BOT.getJDA().getPing());
             lastUpdated = System.currentTimeMillis();