+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 Key: [+] added, [-] removed, [~] modified, [!] important.
 
+## 2019-06-06
++ [~] Refactored all of the audio commands to remove some unnecessary method calls.
++ [~] Added an audio timeout feature which will make Yuuko leave an audio channel after 5 minutes of nothing being played.
+
 ## 2019-05-29
 + [!] Refactored the core GenericMessageController to GenericGuildMessageController to stop Yuuko listening to non-guild messages.
 + [!] Refactored the core GenericMessageReactionController to GenericGuildReactionController to stop Yuuko listening to non-guild reacts.